testing/web-platform/tests/workers/WorkerLocation_search.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerLocation URL decomposition IDL attribute: search </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = './support/WorkerLocation.js';
    var SearchString = "?SearchString";
    var ExpectedResult = SearchString;
    
    var t = async_test("Test Description: WorkerLocation search attribute returns the current query component in the underlying URL.");
                  
    var worker = new Worker(FileName + SearchString);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.search, ExpectedResult);
        t.done();
    });
</script>
</body>
</html>
